+++ packages/sif-common-soknad-ds/src/soknad-step/soknadStepTypes.ts
@@ -1,18 +1,20 @@
-export interface StepConfig<STEPS> {
+export interface StepConfig<STEPS, SøknadRoutes = string> {
     id: string;
     index: number;
     route: string;
     nextStep?: STEPS;
-    nextStepRoute?: string;
+    nextStepRoute?: SøknadRoutes;
     backLinkHref?: string;
+    previousStep?: STEPS;
+    previousStepRoute?: SøknadRoutes;
     previousStepTitleIntlKey?: string;
     pageTitleIntlKey: string;
     stepTitleIntlKey: string;
     nextButtonLabelIntlKey: string;
 }
 
-export interface SoknadStepsConfig<STEPS> {
-    [key: string]: StepConfig<STEPS>;
+export interface SoknadStepsConfig<STEPS, SøknadRoutes = string> {
+    [key: string]: StepConfig<STEPS, SøknadRoutes>;
 }
 
 export enum SoknadApplicationType {