+++ packages/sif-common-soknad-ds/src/soknad-step/soknadStepUtils.ts
@@ -27,22 +27,22 @@ const getRootRoute = (applicationType: SoknadApplicationType): string => {
     return `/${applicationType}/`;
 };
 
-const getStepRoute = <STEPS>(stepId: STEPS, applicationType: SoknadApplicationType): string => {
-    return `${getRootRoute(applicationType)}${stepId}`;
+const getStepRoute = <STEPS, SøknadRoutes>(stepId: STEPS, applicationType: SoknadApplicationType): SøknadRoutes => {
+    return applicationType
+        ? (`${getRootRoute(applicationType)}${stepId}` as SøknadRoutes)
+        : (`${stepId}` as SøknadRoutes);
 };
 
-const getStepsConfig = <STEPS extends string>(
+const getStepsConfig = <STEPS extends string, SøknadRoutes>(
     steps: STEPS[],
     applicationType: SoknadApplicationType
-): SoknadStepsConfig<STEPS> => {
+): SoknadStepsConfig<STEPS, SøknadRoutes> => {
     const numSteps = steps.length;
-    const config: SoknadStepsConfig<STEPS> = {};
+    const config: SoknadStepsConfig<STEPS, SøknadRoutes> = {};
     let idx = 0;
     steps.forEach((stepId) => {
         const nextStep = idx < numSteps - 1 ? steps[idx + 1] : undefined;
-        const nextStepRoute = nextStep ? getStepRoute(nextStep, applicationType) : undefined;
         const prevStepId = idx > 0 ? steps[idx - 1] : undefined;
-
         config[stepId] = {
             id: stepId,
             pageTitleIntlKey: `step.${stepId}.pageTitle`,
@@ -51,17 +51,19 @@ const getStepsConfig = <STEPS extends string>(
             route: getStepRoute(stepId, applicationType),
             index: idx,
             backLinkHref: prevStepId ? getStepRoute(prevStepId, applicationType) : undefined,
+            previousStep: prevStepId,
+            previousStepRoute: prevStepId ? getStepRoute<STEPS, SøknadRoutes>(prevStepId, applicationType) : undefined,
             previousStepTitleIntlKey: prevStepId ? `step.${prevStepId}.pageTitle` : undefined,
             nextStep,
-            nextStepRoute,
+            nextStepRoute: nextStep ? getStepRoute<STEPS, SøknadRoutes>(nextStep, applicationType) : undefined,
         };
         idx++;
     });
     return config;
 };
 
-function getStepIndicatorStepsFromConfig<Steps>(
-    stepsConfig: SoknadStepsConfig<Steps>,
+function getStepIndicatorStepsFromConfig<Steps, SøknadRoutes extends string>(
+    stepsConfig: SoknadStepsConfig<Steps, SøknadRoutes>,
     intl: IntlShape
 ): StepIndicatorStep[] {
     return Object.keys(stepsConfig).map((key) => {