+++ packages/sif-common-soknad-ds/src/soknad-step/step/Step.tsx
@@ -1,13 +1,12 @@
 import { Heading } from '@navikt/ds-react';
 import React from 'react';
+import { useNavigate } from 'react-router-dom';
 import BackLink from '@navikt/sif-common-core-ds/lib/components/back-link/BackLink';
 import Block from '@navikt/sif-common-core-ds/lib/components/block/Block';
 import Page from '@navikt/sif-common-core-ds/lib/components/page/Page';
 import SoknadHeader from '@navikt/sif-common-core-ds/lib/components/soknad-header/SoknadHeader';
 import StepFooter from '@navikt/sif-common-core-ds/lib/components/step-footer/StepFooter';
 import bemHelper from '@navikt/sif-common-core-ds/lib/utils/bemUtils';
-import { FormikValidationErrorSummary } from '@navikt/sif-common-formik-ds/lib';
-import { History } from 'history';
 import StepIndicator, { StepIndicatorStep } from '../step-indicator/StepIndicator';
 import './step.scss';
 
@@ -23,7 +22,7 @@ interface Props {
     previousStepTitle?: string;
     children: React.ReactNode;
     showStepIndicator?: boolean;
-    useValidationErrorSummary?: boolean;
+    validationSummary?: React.ReactNode;
     onBackClick?: () => void;
     topContentRenderer?: () => React.ReactElement<any>;
     onCancel?: () => void;
@@ -38,16 +37,17 @@ function Step({
     backLinkHref,
     steps,
     activeStepId,
-    useValidationErrorSummary,
     onCancel,
     onContinueLater,
     onBackClick,
+    validationSummary,
     cancelOrContinueLaterAriaLabel,
     showStepIndicator = true,
     children,
     previousStepTitle,
 }: Props) {
     const currentStepIndex = steps.findIndex((s) => s.id === activeStepId);
+    const navigate = useNavigate();
     return (
         <Page
             className={bem.block}
@@ -59,7 +59,7 @@ function Step({
                             <SoknadHeader title={bannerTitle} />
                         </>
                     )}
-                    {useValidationErrorSummary !== false && <FormikValidationErrorSummary />}
+                    {validationSummary}
                 </>
             )}>
             {(showStepIndicator || backLinkHref) && (
@@ -69,9 +69,9 @@ function Step({
                             href={backLinkHref}
                             ariaLabel={previousStepTitle}
                             className={bem.element('backLink')}
-                            onClick={(nextHref: string, history: History, event: React.SyntheticEvent) => {
+                            onClick={(nextHref, event) => {
                                 event.preventDefault();
-                                history.push(nextHref);
+                                navigate(nextHref);
                                 if (onBackClick) {
                                     onBackClick();
                                 }