+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @navikt/sif-common-forms-ds
 
+## 3.0.0
+
+### Minor Changes
+
+-   509383d: Pakkeoppdateringer, inkl axios
+
+### Patch Changes
+
+-   Updated dependencies [509383d]
+    -   @navikt/sif-common-core-ds@5.0.0
+
+## 2.0.2
+
+### Patch Changes
+
+-   @navikt/sif-common-core-ds@5.0.0
+
+## 2.0.1
+
+### Patch Changes
+
+-   8533e79: Pakkeoppdateringer
+-   Updated dependencies [8533e79]
+    -   @navikt/sif-common-core-ds@4.4.1
+
+## 2.0.0
+
+### Patch Changes
+
+-   03aeac4: Oppdatere pakker. Ink. major update av react-router-dom.
+-   Updated dependencies [03aeac4]
+    -   @navikt/sif-common-core-ds@4.4.0
+
 ## 1.5.0
 
 ### Minor Changes