+++ packages/sif-common-sentry/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @navikt/sif-common-sentry
 
+## 0.5.0
+
+### Minor Changes
+
+-   509383d: Pakkeoppdateringer, inkl axios
+
+## 0.4.3
+
+### Patch Changes
+
+-   9f5168f: Håndtere undefined headers
+
+## 0.4.2
+
+### Patch Changes
+
+-   8533e79: Pakkeoppdateringer
+
+## 0.4.1
+
+### Patch Changes
+
+-   03aeac4: Oppdatere pakker. Ink. major update av react-router-dom.
+
 ## 0.4.0
 
 ### Minor Changes